file: set permission for backup (security fix)
backup action is a simply copy of the file before edition.
The permission of the backup depends on the default umask.
That is not the best default behaviour.
A backup should be non writable and non executable, and should
be only read by owner, so 0o0400 is expected.
backup_mode property allow to override this new behaviour.